+++ src/components/data-list/index.ts
@@ -1,0 +1,2 @@
+export { MPDataList, MPDataListItem } from './MPDataList';
+export type { MPDataListItemProps, MPDataListProps } from './MPDataList';