+++ src/index.ts
@@ -50,6 +50,7 @@ export * from './components/color-picker';
 export * from './components/combobox';
 export * from './components/command-palette';
 export * from './components/container';
+export * from './components/data-list';
 export * from './components/date-picker';
 export * from './components/date-range-picker';
 export * from './components/date-time-picker';